// client/src/main/cv-viewer/src/app/api/model/cv.ts
/**
 * CV Viewer
 * CV Viewer.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: neilje123@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Education } from './education';
import { Employment } from './employment';
import { Header } from './header';
import { Skills } from './skills';


export interface Cv { 
    name?: string;
    firstName?: string;
    oneLineSummary?: string;
    summary?: string;
    footnote?: string;
    sourceURL?: string;
    interests?: Array<string>;
    image?: string;
    header?: Header;
    education?: Education;
    skills?: Skills;
    employment?: Employment;
}